Heroku

+++ src/main/java/org/generation/BlogPessoal/controller/TemaController.java
@@ -1,6 +1,7 @@
 package org.generation.BlogPessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -17,46 +18,55 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.server.ResponseStatusException;
 
 @RestController
 @RequestMapping("/tema")
-@CrossOrigin
+@CrossOrigin(origins = "*", allowedHeaders = "*")
 public class TemaController {
 
 	@Autowired
 	private TemaRepository repository;
 
-	@GetMapping("/all")
-	public ResponseEntity<List<Tema>> findAll() {
+	@GetMapping
+	public ResponseEntity<List<Tema>> getAll(){
 		return ResponseEntity.ok(repository.findAll());
 	}
 
-	@GetMapping("/id/{id}")
-	public ResponseEntity<Tema> findById(@PathVariable(value = "id") long id) {
-		return repository.findById(id).map(resp -> ResponseEntity.status(200).body(resp))
-				.orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Id não existe"));
+	@GetMapping("/{id}")
+	public ResponseEntity<Tema> getById(@PathVariable long id){
+		return repository.findById(id).map(resp -> ResponseEntity.ok(resp))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
-	@GetMapping("/nome/{nome}")
-	public ResponseEntity<Tema> findByNome(@PathVariable(value = "nome") String nome) {
-		return repository.findByNome(nome).map(resp -> ResponseEntity.status(200).body(resp))
-				.orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Nome não existe"));
+	@GetMapping("/descricao/{descricao}")
+	public ResponseEntity<List<Tema>> getByTitle(@PathVariable String descricao){
+		return ResponseEntity.ok(repository.findAllByDescricaoContainingIgnoreCase(descricao));
 	}
 
-	@PostMapping("/save")
-	public ResponseEntity<Tema> post(@Valid @RequestBody Tema tema) {
-		return ResponseEntity.status(201).body(repository.save(tema));
+	@PostMapping
+	public ResponseEntity<Tema> post(@Valid @RequestBody Tema tema){
+		return ResponseEntity.status(HttpStatus.CREATED)
+				.body(repository.save(tema));
 	}
 
-	@PutMapping("/update")
-	public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema) {
-		return ResponseEntity.status(HttpStatus.OK).body(repository.save(tema));
+	@PutMapping
+	public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema){
+		return repository.findById(tema.getId())
+				.map(resp -> ResponseEntity.status(HttpStatus.CREATED).body(repository.save(tema)))
+				.orElse(ResponseEntity.status(HttpStatus.BAD_REQUEST).build());
 	}
 
-	@DeleteMapping("/id/{id}")
-	public void delete(@PathVariable(value = "id") long id) {
+	@ResponseStatus(HttpStatus.NO_CONTENT)
+	@DeleteMapping("/{id}")
+	public void delete(@PathVariable long id) {
+		Optional<Tema> tema = repository.findById(id);
+
+		if(tema.isEmpty())
+			throw new ResponseStatusException(HttpStatus.BAD_REQUEST);
+
 		repository.deleteById(id);
 	}
